+++ clients/vscode-hlasmplugin/CHANGELOG.md
@@ -6,6 +6,9 @@
 - Provide more details about machine instructions in hover texts
 - Faded preprocessor statements
 
+#### Fixed
+- Large macro documentation is not highlighted correctly in hover texts
+
 ## [1.6.0](https://github.com/eclipse/che-che4z-lsp-for-hlasm/compare/1.5.0...1.6.0) (2023-01-18)
 
 #### Added